requirements.txt
Django==4.0.2
djangorestframework==3.13.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
celery==5.2.3
redis==4.1.4
psycopg2==2.9.2